Reorganized the views of each app separately, rewritten the paths in the urls.py file, recreated the migrations
