+++ Django_Frolov/settings.py
@@ -37,7 +37,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'students'
+    'students',
+    'teachers',
+    'group',
 ]
 
 MIDDLEWARE = [
