+++ group/views.py
@@ -1,3 +1,12 @@
-from django.shortcuts import render
+from django.http import HttpResponse
 
-# Create your views here.
+from .models import Group
+
+
+def groups(request):
+    groups_list = Group.objects.all()
+    output = ''.join(
+        [f"<p>Group name: {group.group_name}, Faculty name: {group.faculty_name} - "
+         f"group consists of {group.nos} students;</p>" for group in groups_list]
+        )
+    return HttpResponse(output)
