+++ students/admin.py
@@ -1,7 +1,5 @@
 from django.contrib import admin
 
-from .models import Student, Group, Teacher
+from .models import Student
 
 admin.site.register(Student)
-admin.site.register(Group)
-admin.site.register(Teacher)
